+++ src/main/java/com/nix/repository/ChapterRepository.java
@@ -9,15 +9,15 @@ import org.springframework.data.jpa.repository.Query;
 import com.nix.models.Book;
 import com.nix.models.Chapter;
 
-public interface ChapterRepository extends JpaRepository<Chapter, Integer> {
+public interface ChapterRepository extends JpaRepository<Chapter, Long> {
 	public Chapter findByTitle(String title);
 
-	public List<Chapter> findByBookId(Integer bookId);
+	public List<Chapter> findByBookId(Long bookId);
 
 	@Query("select c from Chapter c where c.isDraft=false and c.book.id=:bookId")
-	public List<Chapter> findNotDraftedChaptersByBookId(Integer bookId);
+	public List<Chapter> findNotDraftedChaptersByBookId(Long bookId);
 
-	Optional<Chapter> findByIdAndIsLocked(Integer id, boolean isLocked);
+	Optional<Chapter> findByIdAndIsLocked(Long id, boolean isLocked);
 
 	@Query("select c.book b from Chapter c order by c.uploadDate DESC limit 5 ")
 	public List<Book> findTopByOrderByUploadDateDesc();
